+++ web/src/components/planner/Planner.vue
@@ -5,6 +5,7 @@
   <planner-too-many-factories-open :factories="getFactories()" @hide-all="showHideAll('hide')" />
 
   <building-group-tutorial />
+  <awesome-sink-tutorial />
   <div class="planner-container" :class="{ 'full-width': plannerOptions.fullWidth }">
     <!-- Navigation Drawer for Mobile -->
     <Teleport v-if="navigationReady" defer to="#navigationDrawer">
@@ -149,6 +150,7 @@
   import { FactoryGroupSection } from '@/utils/factory-management/factory-groups'
   import eventBus from '@/utils/eventBus'
   import BuildingGroupTutorial from '@/components/planner/products/BuildingGroupTutorial.vue'
+  import AwesomeSinkTutorial from '@/components/planner/AwesomeSinkTutorial.vue'
   import PlannerGroupBand from '@/components/planner/groups/PlannerGroupBand.vue'
   import DimensionalDepot from '@/components/planner/DimensionalDepot.vue'
   import { groupColorVars } from '@/utils/colors'